backend/requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
Deprecated==1.2.18
dill==0.3.8
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.8
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-ollama==0.2.3
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.16
loky==3.5.6
lxml==6.0.1
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
ollama==0.5.3
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.1
packaging==24.2
pandas==2.3.2
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
